+++ src/Validator/Collection/ContainsValidator.php
@@ -6,7 +6,7 @@ namespace ExtendsSoftware\ExaPHP\Validator\Collection;
 
 use ExtendsSoftware\ExaPHP\Validator\AbstractValidator;
 use ExtendsSoftware\ExaPHP\Validator\Exception\TemplateNotFound;
-use ExtendsSoftware\ExaPHP\Validator\Result\Container\ContainerResult;
+use ExtendsSoftware\ExaPHP\Validator\Result\Container\Array\ArrayContainerResult;
 use ExtendsSoftware\ExaPHP\Validator\Result\ResultInterface;
 use ExtendsSoftware\ExaPHP\Validator\Type\IterableValidator;
 use ExtendsSoftware\ExaPHP\Validator\ValidatorInterface;
@@ -41,11 +41,10 @@ class ContainsValidator extends AbstractValidator
             return $result;
         }
 
-        $container = new ContainerResult();
-        foreach ($value as $index => $inner) {
-            $container->addResult(
+        $container = new ArrayContainerResult();
+        foreach ($value as $inner) {
+            $container->addItem(
                 $this->validator->validate($inner, $context),
-                (string)$index
             );
         }