+++ src/Validator/Result/ResultInterface.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace ExtendsSoftware\ExaPHP\Validator\Result;
 
+use ExtendsSoftware\ExaPHP\Validator\Result\Exception\ResultNotValid;
 use JsonSerializable;
 
 interface ResultInterface extends JsonSerializable
@@ -14,4 +15,12 @@ interface ResultInterface extends JsonSerializable
      * @return bool
      */
     public function isValid(): bool;
+
+    /**
+     * Get the validated valid value.
+     *
+     * @return mixed
+     * @throws ResultNotValid
+     */
+    public function getValue(): mixed;
 }